Refactoring : Data Clump
